+++ identity_stardust/examples/ex0_create_did.rs
